Keep Opus CELT input bridge float-width

+++ celt/celt_encode.go
@@ -6,7 +6,7 @@ package celt
 // Package-level stateless helpers for simple API.
 
 // Encode encodes mono PCM samples to a CELT packet.
-// pcm: float64 samples at 48kHz
+// pcm: float32 samples at 48kHz
 // frameSize: 120, 240, 480, or 960 samples
 // Returns: encoded Opus CELT packet bytes
 //
@@ -14,13 +14,13 @@ package celt
 // use NewEncoder() and call EncodeFrame() directly.
 //
 // Reference: RFC 6716 Section 4.3
-func Encode(pcm []float64, frameSize int) ([]byte, error) {
+func Encode(pcm []float32, frameSize int) ([]byte, error) {
 	enc := NewEncoder(1)
 	return enc.EncodeFrame(pcm, frameSize)
 }
 
 // EncodeStereo encodes stereo PCM samples to a CELT packet.
-// pcm: interleaved L/R float64 samples at 48kHz
+// pcm: interleaved L/R float32 samples at 48kHz
 // frameSize: 120, 240, 480, or 960 samples per channel
 // Returns: encoded Opus CELT packet bytes
 //
@@ -30,14 +30,14 @@ func Encode(pcm []float64, frameSize int) ([]byte, error) {
 // This uses mid-side stereo encoding (dual_stereo=0, intensity disabled).
 //
 // Reference: RFC 6716 Section 4.3
-func EncodeStereo(pcm []float64, frameSize int) ([]byte, error) {
+func EncodeStereo(pcm []float32, frameSize int) ([]byte, error) {
 	enc := NewEncoder(2)
 	return enc.EncodeFrame(pcm, frameSize)
 }
 
 // EncodeWithEncoder encodes mono PCM using the provided encoder.
 // Allows stateful encoding with custom encoder instances.
-func EncodeWithEncoder(enc *Encoder, pcm []float64, frameSize int) ([]byte, error) {
+func EncodeWithEncoder(enc *Encoder, pcm []float32, frameSize int) ([]byte, error) {
 	if enc == nil {
 		return nil, ErrEncodingFailed
 	}
@@ -46,7 +46,7 @@ func EncodeWithEncoder(enc *Encoder, pcm []float64, frameSize int) ([]byte, erro
 
 // EncodeStereoWithEncoder encodes stereo PCM using the provided encoder.
 // Allows stateful encoding with custom encoder instances.
-func EncodeStereoWithEncoder(enc *Encoder, pcm []float64, frameSize int) ([]byte, error) {
+func EncodeStereoWithEncoder(enc *Encoder, pcm []float32, frameSize int) ([]byte, error) {
 	if enc == nil {
 		return nil, ErrEncodingFailed
 	}
@@ -61,7 +61,7 @@ func EncodeStereoWithEncoder(enc *Encoder, pcm []float64, frameSize int) ([]byte
 // pcmFrames: slice of PCM frames, each with frameSize samples
 // frameSize: samples per frame (must be same for all frames)
 // Returns: slice of encoded packets
-func EncodeFrames(pcmFrames [][]float64, frameSize int) ([][]byte, error) {
+func EncodeFrames(pcmFrames [][]float32, frameSize int) ([][]byte, error) {
 	if len(pcmFrames) == 0 {
 		return nil, nil
 	}
@@ -84,7 +84,7 @@ func EncodeFrames(pcmFrames [][]float64, frameSize int) ([][]byte, error) {
 // pcmFrames: slice of interleaved stereo PCM frames
 // frameSize: samples per frame per channel
 // Returns: slice of encoded packets
-func EncodeStereoFrames(pcmFrames [][]float64, frameSize int) ([][]byte, error) {
+func EncodeStereoFrames(pcmFrames [][]float32, frameSize int) ([][]byte, error) {
 	if len(pcmFrames) == 0 {
 		return nil, nil
 	}
@@ -106,7 +106,7 @@ func EncodeStereoFrames(pcmFrames [][]float64, frameSize int) ([][]byte, error) 
 // EncodeSilence encodes a silent frame of the given size.
 // Useful for generating comfort noise or filler packets.
 func EncodeSilence(frameSize int, channels int) ([]byte, error) {
-	pcm := make([]float64, frameSize*channels)
+	pcm := make([]float32, frameSize*channels)
 
 	if channels == 1 {
 		return Encode(pcm, frameSize)
